requirements.txt
aiohttp>=3.8.0
rich>=13.0.0
typer>=0.9.0
dnspython>=2.4.0
asyncio>=3.4.3
aiofiles>=23.2.1